# backend/docs/contracts/generated-openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/sms/campaigns": {
      "post": {
        "operationId": "SmsCampaignController_createCampaign",
        "summary": "Create SMS campaign",
        "parameters": [],
        "responses": {
          "201": {
            "description": "Campaign created"
          }
        },
        "tags": ["SMS Campaigns"]
      },
      "get": {
        "operationId": "SmsCampaignController_getCampaigns",
        "summary": "Get all campaigns",
        "parameters": [
          {
            "name": "status",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "createdBy",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Campaigns retrieved"
          }
        },
        "tags": ["SMS Campaigns"]
      }
    },
    "/sms/campaigns/{id}": {
      "get": {
        "operationId": "SmsCampaignController_getCampaign",
        "summary": "Get campaign by ID",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Campaign retrieved"
          }
        },
        "tags": ["SMS Campaigns"]
      },
      "delete": {
        "operationId": "SmsCampaignController_deleteCampaign",
        "summary": "Delete campaign",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Campaign deleted"
          }
        },
        "tags": ["SMS Campaigns"]
      }
    },
...
    "/admin/sms/campaigns/{id}": {
      "get": {
        "operationId": "SmsAdminController_getCampaign",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        }
      },
      "delete": {
        "operationId": "SmsAdminController_deleteCampaign",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        }
      }
    },
...
    "/sms/campaigns/{id}/status": {
      "patch": {
        "operationId": "SmsCampaignController_updateCampaignStatus",
        "summary": "Update campaign status",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Campaign status updated"
          }
        },
        "tags": ["SMS Campaigns"]
      }
    },
    "/sms/campaigns/{id}/stats": {
      "patch": {
        "operationId": "SmsCampaignController_updateCampaignStats",
        "summary": "Update campaign statistics",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Campaign stats updated"
          }
        },
        "tags": ["SMS Campaigns"]
      }
    },
...
    "/media/upload": {
      "post": {
        "operationId": "MediaController_upload",
        "summary": "Upload media file",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": ["Media"],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/media/{id}": {
      "get": {
        "operationId": "MediaController_findOne",
        "summary": "Get media by ID",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": ["Media"]
      },
      "delete": {
        "operationId": "MediaController_remove",
        "summary": "Delete media",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": ["Media"],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
    "/experts": {
      "get": {
        "operationId": "KnowledgeController_findAll",
        "summary": "List experts",
        "description": "Returns a list of experts with optional filtering by department and legend status. Results can be limited to a specific number of items.",
        "parameters": [
          {
            "name": "departmentId",
            "required": false,
            "in": "query",
            "description": "Filter by department ID",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "isLegend",
            "required": false,
            "in": "query",
            "description": "Filter for legends only (true/false)",
            "schema": {
              "example": "false",
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Maximum number of experts to return",
            "schema": {
              "example": "20",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Experts list retrieved successfully"
          }
        },
        "tags": ["Knowledge"]
      }
    },
...
    "/experts/skills/{slug}": {
      "get": {
        "operationId": "KnowledgeController_getSkillBySlug",
        "summary": "Get skill by slug",
        "description": "Returns detailed information about a specific skill by its slug identifier.",
        "parameters": [
          {
            "name": "slug",
            "required": true,
            "in": "path",
            "description": "Skill slug (URL-friendly identifier)",
            "schema": {
              "example": "video-editing",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Skill details retrieved successfully"
          },
          "404": {
            "description": "Skill not found"
          }
        },
        "tags": ["Knowledge"]
      }
    },
    "/experts/{id}": {
      "get": {
        "operationId": "KnowledgeController_findOne",
        "summary": "Get expert by ID",
        "description": "Returns detailed profile information for a specific expert by their ID.",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Expert ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Expert profile retrieved successfully"
          },
          "404": {
            "description": "Expert not found"
          }
        },
        "tags": ["Knowledge"]
      },
      "put": {
        "operationId": "KnowledgeController_update",
        "summary": "Update expert profile",
        "description": "Updates an expert profile information. Requires authentication.",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Expert ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Expert profile updated successfully"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Expert not found"
          }
        },
        "tags": ["Knowledge"],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
    "/notifications/{id}/read": {
      "post": {
        "operationId": "NotificationController_markAsRead",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        }
      }
    },
...
    "/pdf/generate": {
      "post": {
        "operationId": "PdfGeneratorController_generatePdf",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GeneratePdfDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        }
      }
    },
...
  },
  "info": {
    "title": "IRIB DWP API",
    "description": "IRIB East Azerbaijan Digital Workplace Platform API",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "GeneratePdfDto": {
        "type": "object",
        "properties": {}
      },
...
    }
  }
}