# backend/src/modules/media/media.controller.ts
import {
  Controller,
  Post,
  Get,
  Delete,
  Param,
  UploadedFile,
  UseInterceptors,
  UseGuards,
  Request,
} from '@nestjs/common'
import { FileInterceptor } from '@nestjs/platform-express'
import { ApiTags, ApiOperation, ApiBearerAuth, ApiConsumes } from '@nestjs/swagger'
import { MediaService } from './media.service'
import { JwtAuthGuard } from '../iam/jwt-auth.guard'

@ApiTags('Media')
@Controller('media')
export class MediaController {
  constructor(private readonly mediaService: MediaService) {}

  @Post('upload')
  @UseGuards(JwtAuthGuard)
  @UseInterceptors(FileInterceptor('file'))
  @ApiBearerAuth()
  @ApiConsumes('multipart/form-data')
  @ApiOperation({ summary: 'Upload media file' })
  async upload(@UploadedFile() file: Express.Multer.File, @Request() req: any) {
    return this.mediaService.upload(file, req.user.sub)
  }

  @Get(':id')
  @ApiOperation({ summary: 'Get media by ID' })
  async findOne(@Param('id') id: string) {
    return this.mediaService.findOne(id)
  }

  @Delete(':id')
  @UseGuards(JwtAuthGuard)
  @ApiBearerAuth()
  @ApiOperation({ summary: 'Delete media' })
  async remove(@Param('id') id: string) {
    return this.mediaService.delete(id)
  }

  @Get('stats/upload')
  @ApiOperation({ summary: 'Get upload statistics' })
  async getStats() {
    return this.mediaService.getUploadStats()
  }
}